[REFACTOR] remove navbar on login
