+++ __test__/semuaBarang/semuaBarang.test.tsx
@@ -1,6 +1,6 @@
 import '@testing-library/jest-dom'
 import { render, screen } from "@testing-library/react"
-import SemuaBarang from "../../src/app/semuaBarang/page"
+import SemuaBarang from "../../src/app/(withNavbar)/semuaBarang/page"
 
 describe("Semua Barang Page", () => {
   beforeEach(() => {
